requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
chromadb==0.4.22
sentence-transformers==2.3.1
python-dotenv==1.0.0
google-generativeai==0.8.3
langchain==0.3.1
langchain-google-genai==2.1.10
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0